+++ order-service/order-domain/order-application-service/src/main/java/me/jincrates/ecommerce/order/outbox/model/OrderPaymentEventPayload.java
@@ -1,7 +1,7 @@
 package me.jincrates.ecommerce.order.outbox.model;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.time.ZonedDateTime;
 
 public record OrderPaymentEventPayload(
@@ -10,7 +10,7 @@ public record OrderPaymentEventPayload(
     @JsonProperty
     String customerId,
     @JsonProperty
-    BigDecimal price,
+    BigInteger price,
     @JsonProperty
     ZonedDateTime createdAt,
     @JsonProperty
